requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
requests
selenium
python-dotenv
schedule
